# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Felt API Doc",
        "description": "Felt API Documentation",
        "contact": {
            "email": "rejohnbd@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customer/customer-vehicles": {
            "get": {
                "tags": [
                    "customer-vehicles"
                ],
                "summary": "Returns all customer vehicle ",
                "description": "",
                "operationId": "customer-vehicles",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "description": "shortList",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/customer-vehicles/id": {
            "get": {
                "tags": [
                    "customer-single-vehicle"
                ],
                "summary": "Get customer single vehicle ",
                "description": "",
                "operationId": "customer-single-vehicle",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/vehicle-update": {
            "post": {
                "tags": [
                    "vehicle-update-purpose"
                ],
                "summary": "Update Vehicle Purpose",
                "description": "Update Vehicle Purpose.",
                "operationId": "vehicle-update-purpose",
                "parameters": [
                    {
                        "name": "vehicle_id",
                        "in": "path",
                        "description": "Vehicle Id",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "vehicle_purpose",
                        "in": "path",
                        "description": "Vehicle Purpose",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fuel_capacity",
                        "in": "path",
                        "description": "Fuel Congestion",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "vehicle_kpl",
                        "in": "path",
                        "description": "Milage",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "driver_id",
                        "in": "path",
                        "description": "Driver Id",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid user supplied"
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/customer/customer-vehicles-speed-limitation/id": {
            "get": {
                "tags": [
                    "customer-vehicles-speed-limitation"
                ],
                "summary": "Get customer vehicle speed info",
                "description": "",
                "operationId": "customer-vehicles-speed-limitation",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer token": []
                    }
                ]
            }
        },
        "/api/customer/customer-vehicles-speed-limitation": {
            "post": {
                "tags": [
                    "customer-vehicles-speed-limitation-update"
                ],
                "summary": "Update customer vehicle speed info",
                "description": "",
                "operationId": "customer-vehicles-speed-limitation-update",
                "parameters": [
                    {
                        "name": "vehicle_id[]",
                        "in": "path",
                        "description": "Vehicle Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "speed_limitation[]",
                        "in": "path",
                        "description": "Speed Limitation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer token": []
                    }
                ]
            }
        },
        "/api/customer/ccustomer-vehicles-alert-setting/id": {
            "get": {
                "tags": [
                    "customer-vehicles-alert-setting"
                ],
                "summary": "Get Customer Vehicle Alert Info By Id",
                "description": "",
                "operationId": "customer-vehicles-alert-setting",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/customer/customer-vehicles-speed-alert-setting": {
            "get": {
                "tags": [
                    "customer-vehicles-speed-alert-setting"
                ],
                "summary": "Get Customer Vehicle Speed & Alert Info",
                "description": "",
                "operationId": "customer-vehicles-speed-alert-setting",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "additionalProperties": {
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}